chore: extract keystore metadata, add migration
